requirements.txt
Django==5.2.1
django-debug-toolbar==4.2.0
Pillow==10.0.0
reportlab==4.0.4
djangorestframework==3.14.0
django-filter==23.2